navbar & routes created v2
